ISPN-6152 Move the jdk8backported package down one level
* Removed the package entirely
* Added upgrade documentation
