+++ core/src/main/java/org/infinispan/container/entries/PrimitiveEntrySizeCalculator.java
@@ -2,7 +2,7 @@ package org.infinispan.container.entries;
 
 import java.lang.reflect.Array;
 
-import org.infinispan.commons.util.concurrent.jdk8backported.AbstractEntrySizeCalculatorHelper;
+import org.infinispan.commons.util.AbstractEntrySizeCalculatorHelper;
 
 import sun.misc.Unsafe;
